# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AM Trading PLC API",
        "description": "Inventory & sales management API for AM Trading PLC. All endpoints require Bearer token authentication except `POST /auth/login`.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/sales": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "List sales",
                "operationId": "a2e981e47fac573d9961105517d022df",
                "parameters": [
                    {
                        "name": "filter[customer_id]",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[branch_id]",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[payment_status]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "paid",
                                "partial",
                                "credit",
                                "unpaid"
                            ]
                        }
                    },
                    {
                        "name": "filter[date_from]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "filter[date_to]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated sale list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Create a sale",
                "operationId": "678fe5b736161e60f22fb2cc5ae9e827",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_id",
                                    "warehouse_id",
                                    "sale_date",
                                    "payment_method",
                                    "items"
                                ],
                                "properties": {
                                    "branch_id": {
                                        "type": "integer"
                                    },
                                    "warehouse_id": {
                                        "type": "integer"
                                    },
                                    "customer_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "sale_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "bank_transfer",
                                            "telebirr",
                                            "credit_card",
                                            "check",
                                            "full_credit",
                                            "credit_advance"
                                        ]
                                    },
                                    "advance_amount": {
                                        "type": "number"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "item_id",
                                                "quantity",
                                                "unit_price"
                                            ],
                                            "properties": {
                                                "item_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                },
                                                "unit_price": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sale created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Sale"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sales/{id}": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Get a sale",
                "operationId": "b0c4d552c576c27953130c56d104a408",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sale detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Sale"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sales"
                ],
                "summary": "Delete a sale",
                "operationId": "db554661d18c032b8ed4e5a438fdd379",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/stocks": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "List stock levels",
                "operationId": "adf2f7bee2bcea557a9458d225f20796",
                "parameters": [
                    {
                        "name": "filter[branch_id]",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[warehouse_id]",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[item_id]",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[below_reorder]",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stock list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/stocks/{id}": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Get a stock record",
                "operationId": "4ce763191d9e8202dc5af23979987b24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stock detail"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Branch": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Main Branch"
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Customer": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "branch": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Branch"
                            }
                        ],
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "SaleItem": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "item_id": {
                        "type": "integer"
                    },
                    "item_name": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float"
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "Sale": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "reference_no": {
                        "type": "string",
                        "example": "SO-2024-0001"
                    },
                    "customer": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Customer"
                            }
                        ],
                        "nullable": true
                    },
                    "branch": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Branch"
                            }
                        ],
                        "nullable": true
                    },
                    "sale_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "payment_method": {
                        "type": "string",
                        "enum": [
                            "cash",
                            "bank_transfer",
                            "telebirr",
                            "credit_card",
                            "check",
                            "full_credit",
                            "credit_advance"
                        ]
                    },
                    "payment_status": {
                        "type": "string",
                        "enum": [
                            "paid",
                            "partial",
                            "credit",
                            "unpaid"
                        ]
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "paid_amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "advance_amount": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "due_amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "note": {
                        "type": "string",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SaleItem"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token — obtain via POST /auth/login",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}